(maint) Pass in a custom tmpdir to install_puppetlabs_dev_repos
Because the install_puppetlabs_dev_repo uses the same tmpdir under the
hood to prepare repos, if the contents of the repos are not the same,
they will end up causing package install failures, as was seen today
when a new Packages.bz2 file caused the second repo to be corrupted.
This commit passes in a custom tmpdir to both calls to avoid this
problem.