(SERVER-1490) Add crl reload workaround for PE
This commit adds a PE-specific workaround to the `reload_updated_crl`
acceptance test.  To compensate for the lack of shadowing of the hostcrl
to cacrl file in PE (documented in SERVER-911), the test manually copies
the file (for PE only) so that the updated CRL is in place after the
server is restarted.