(SERVER-1570) ExecutionStub should raise error on failonfail
Prior to this commit, Puppet Server's implementation of `ExecutionStub`
would ignore the `:failonfail` option.  The documented behavior is to
raise a `Puppet::ExecutionFailure` if the command returns a non-zero
exit code.

This commit refactors the ExecutionStub code a bit to allow the
options hash to be passed in (which also required some refactoring
of the tests).  Then, if `:failonfail` is set, we raise the appropriate
error when the exit code is non-zero.