(SERVER-1570) Clean up logic for parsing command
This commit just cleans up some of the logic from the previous
commit, where we were parsing the command string/array before
we called the private `execute` method, which forced us to
pass the original command string in as an argument.  Now we
do that parsing inside of the private method, making the signatures
and the test code a little cleaner.