(SERVER-1571) ensure environment passed to clj shellutils
A recent refactor resulted in a subtle change in behavior in the
clojure shellutils wrapper, wherein if you did not pass an
explicit list of environment variables to the command, you would
end up with an empty map of environment variables for the shell
command.  Prior to that refactor, you would instead have inherited
all of the environment variables from the java process, which is
the desired behavior.  This change broke some acceptance tests which
were relying on having access to environment variables such as PATH.

This commit changes the environment variable behavior back to the
previous setup, where the java environment variables will be passed
through to the shell command unless you explicitly override the
environment.