(MAINT) Loosen SSL_connect error check in reload_updated_crl test
This commit loosens the string which is checked in the
reload_updated_crl test for an SSL connection failure on a run made
from a revoked agent.  The test previously checked for 'SSL_connect
SYSCALL returned=5' but the 'SYSCALL returned=5' part may or may not
appear depending upon how/when the agent detects the handshake
failure.  Occasionally, instead, it just appears as a 'Connection
reset by peer - SSL_connect' message.  This commit just loosens the
text to 'SSL_connect', which should be good enough to determine that
the connection failed even if the exact nature of the failure is a
bit less precise.