(SERVER-1601) Set umask for gem subcommand to 0022
The bump to EZBake 1.0+ causes cli subcommands to run with a default
umask of 027.  Previously, gems installed via the puppetserver gem
subcommand would have the world read bits for their gemspec files masked
out by this umask even though the gemspec files could be owned by the
root user.  This could cause puppetserver service, running as user
'puppet', to fail to load gems properly.

This commit sets the umask for the gem subcommand to 0022.  This should
allow gemspec files to be world-readable and, therefore, loadable by the
puppetserver service.