(SERVER-797) Allow jvm heap size to be configurable for dev/test
This commit adds options for configuring the min/max heap size which is
used when running lein project commands.  This allows for the heap size
to be bumped up to a significantly higher value for local development
purposes, with the benefit of significantly reducing the total run time
for the test suite.  An alternate min/max heap size can be configured
via either an environment variable, PUPPETSERVER_HEAP_SIZE, or a
profiles.clj setting, user.puppetserver-heap-size.