(SERVER-797) Bump up Travis heap size for faster test runs
This commit bumps up the heap size when running Clojure tests in Travis
to 6G, in hopes of improving test run performance.  This also moves from
the Travis container-based infrastructure back to the "standard"
infrastructure, by virtue of setting "sudo: required" in the .travis.yml
file in order to have more memory available for bumping up the heap.