(PE-18344) Set http client metric registry if metrics service is present
Use the Trapperkeeper optional service dependency functionality to get a
metric registry if the metrics service is present, and pass it and the
`server-id` through to the http client that puppet server uses, so that the
http client will register metrics. If no metrics service is present, the http
client will not get passed a metric registry or server-id and will not
register metrics.