(SERVER-797) Create mock jruby-pool-manager-service and related helpers
This commit includes a new mock `jruby-pool-manager-service` for
testing, with a `create-pool` implementation which creates a pool of
fake JRuby ScriptingContainers.  The commit also includes variations of
the `with-puppetserver-running...` functions which utilize the mock
`jruby-pool-manager-service`.

The fake `ScriptingContainer`s are reified from a new
`ScriptingContainer` interface exposed from the jruby-utils library,
which requires bumping the `jruby-utils` dependency to a newer version,
0.4.1.  The unused ScriptingContainer Java class in the puppetserver
project has also been removed.