(SERVER-797) Modify http_client_tests to use jruby-utils helpers
This commit modified the http_client_test tests to use public helpers in
jruby-utils for creating a ScriptingContainer.  Because the default
ScriptingContainer instances which are created do not provide PERSISTENT
LocalVariableBehavior, which the tests had been depending upon, the
tests have also been modified to use global variables instead of local
ones in Ruby.