(SERVER-797) Reuse a ScriptingContainer throughout the http-client tests
Previously, each of the http-client tests created a new
ScriptingContainer.  This commit adds a test fixture to the http-client
tests which allows a Ruby ScriptingContainer to be created one time and
reused across all of the tests run for the namespace.  These changes
involve cleaning up http-client objects between tests to avoid
contamination, previously done just by throwing the entire
ScriptingContainer away.  Locally, this reduced the run time for the
namespace from about 51 to under 4 seconds.