(SERVER-797) Rework env-class-int-tests to use mock-jruby-pool-manager
Previously, the environment-classes-int-tests used a with-redef for the
`create-pool-instance` function in order to mock the creation of a
JRubyPuppet instance.  This commit reworks the
environment-classes-int-tests to use the mock-jruby-pool-manager service
instead, eliminating the use of the with-redef.  This commit allows a
custom `mock-jruby-puppet-fn` to be passed into the
`mock-jruby-pool-manager` so that individual tests like the
environment-classes-int-tests can customize the behavior of JRubyPuppet
instances.