(SERVER-797) Removed need for temp file from mock jruby dev-bootstrap
Previously, the dev bootstrap with mock jruby pool manager service
helper constructed the list of services to bootstrap by making a temp
file copy of the original dev bootstrap config file and replacing the
line for the real jruby-pool-manager-service with the mock one.  This
commit changes the helper to load the services from the
dev/bootstrap.cfg file and filter out the jruby-pool-manager-service
instance in memory - avoiding the need for a temp file.