(SERVER-797) Add docstrings to mock macros, first pass
This commit adds a required docstring argument to both of the new
JRuby mock macros, to enforce that people need to make a few notes
about why it's safe to use JRuby mocking for a given test.

The commit also makes a first pass over all the uses of the new
macros and adds appropriate docstrings to the ones that seem safe,
and `MOCK.TODO` comments for the ones that warrant further
discussion.