(SERVER-797) Coerce symbols to strings from JRubyPuppet.getSetting()
Previously, if the JRubyPuppet.getSetting() method were to obtain a
value for a setting of type `org.jruby.RubySymbol`, the value would be
passed on through to the Clojure layer.  Where we'd like to avoid having
any JRuby-typed objects leak out of the Ruby layer, this commit coerces
a Ruby `Symbol` return value to a `String`.  For other return types, the
original object is still returned through the interop layer.  This
change only currently affects the `data_binding_terminus` setting -
first plumbed through the puppet-server-config-service for PE-17071.