(SERVER-797) Add assertions that mock matches real JRubyPuppet settings
This commit reworks the `test-puppet-config-values` unit test to assert
that the names and values of the mock JRubyPuppet settings, used by the
mock-jruby-pool-manager-service, match those of the 'real' settings that
the puppet-server-config-service uses with JRubyPuppet instances that
call into the core Ruby Puppet settings code.

Previously, the mock JRubyPuppet.getSetting() method would return nil if
the requested setting were not found.  For this commit, the getSetting()
method will instead throw an Exception if the requested setting is not
found - a further indication of the need to add a definition for any new
setting to the mock settings map.