(SERVER-797) Remove data_binding_terminus plumbing and symbol coercion
This commit removes `data_binding_terminus` from the list of settings
that the puppet-server-config-service requests from JRubyPuppet.  This
is being done because the work done to add this in PE-17071 turned out
not to be needed and not having it present makes the test refactoring
done elsewhere for SERVER-797 a bit simpler.

Corresponding work in the master.rb class to convert a `Symbol` data
type to a `String` for a `JRubyPuppet.getSetting` call was also
removed.  This code was no longer necessary since no settings other than
'data_binding_terminus' had a data type of Symbol.