(SERVER-797) Added a with-puppetserver-running-with-mock-jruby-puppet fn
This commit adds a variant of the `with-puppetserver-running` macros
which allows a mock JRubyPuppet creator function to be supplied.  Along
with the introduction of this function, existing tests which had been
relying implicitly on the default behavior of the `.handleRequest`
function in the default mock JRubyPuppet creator function to return a
200 (OK) response now supply their own definition of the
`.handleRequest` function.  This allows those tests to more
intentionally validate the contents of a request as it would be passed
through to the Ruby interop layer.