(SERVER-797) Change test-503-app-shut-down to stop app even on error
Previously, if the `test-503-when-app-shuts-down` test encounters an
error after an app has been started but before it is shut down, the
Jetty webserver that the test spawned would keep on running -
potentially causing many tests downstream to fail as well.  This commit
adds some cleanup logic which attempts to stop the all services
(including the Jetty webserver) in the event that an unexpected error
causes the test to fail before all assertions are made.