(MAINT) Explicitly set GEM_HOME/PATH and avoid extra/dev gem install
Previously, the GEM_HOME and GEM_PATH variables for installing the
bundler gem from the Rakefile for JRuby spec tests were not set.
This could cause gem content from the default system Ruby to be
pulled in during installation, leading to some odd error messages
at best or possibly failure at worst.  This commit sets the GEM_HOME and
GEM_PATH variables to the local ./vendor directory to avoid potential
collisions with a system MRI Ruby, just as is done for the later bundle
install of gems from the Puppet submodule's Gemfile for the spec run.

This commit also avoids installing the 'extra' and 'development' gems
from the Puppet Gemfile since these are unnecessary for running the
JRuby spec tests.