(SERVER-797) Add a mock-jruby-puppet-fn with handler args helper
This commit adds a new function which can be called by tests to return a
mock-jruby-puppet creator function.  The response to any .handleRequest()
call made to a JRubyInstance created by the function includes parameters
supplied to the new creator function.  Several of the existing tests
were also modified to use this new helper in lieu of calling the
lower-level `create-mock-jruby-puppet` function.