(MAINT) Bump puppet submodule to 1ea537
Previously, JRuby spec tests were failing because of an inability to
install the public_suffix gem.  When the Puppet Gemfile was bundle
installed, version 2.5.0 of the addressable gem tried to pull in the
public_suffix gem, which requires Ruby 2.  JRuby spec tests, however,
run under MRI 1.9.3 compatibility mode, causing the install to fail.

This commit bumps the Puppet submodule up to 1ea537 and corresponding
puppet-agent package pin to 549d028.  This is being done in order to
pick up a change to the Puppet Gemfile, which pins the version of the
addressable gem to < 2.5.0.  Version 2.4.0 of the addressable gem can be
installed and used properly under MRI 1.9.3 compatibility mode.