(SERVER-1580) Switch to `dev_tools` namespace for "reloaded" fns
This commit is the first in a series of commits to try to bring
the Puppet Server repo into alignment with the Clojure dev environment
best practices we recently established on the mailing list.  (Results
can be found [here](https://confluence.puppetlabs.com/pages/viewpage.action?pageId=105578582).)

Specifically, this commit moves all of the "reloaded" workflow functions
into the `dev-tools` namespace, and makes that the default REPL
namespace.  It also introduces a `get-config` fn that will be used by
`go` when bootstrapping the application; a subsequent commit will
make it easy for users to override this.

The commit also simplifies some of the logic for loading the default
configuration and bootstrapped services list.