(SERVER-1580) Add puppet.conf initialization to `dev-setup`
This commit adds some logic to the `dev-setup` to initialize the
`puppet.conf` file according to the recent conversations about
Clojure dev environments.  Specifically, the script will set
the certname for Puppet Server's use; it will honor the value of the
MASTERHOST environment variable if it is defined, and will otherwise
default to `localhost`.