(MAINT) Skip agent_run_before_ssl_files_inited test on PE
Previously, the `master_starts_if_agent_run_before_ssl_files_inited`
acceptance test was configured to run both for OSS and PE Puppet Server.
The test nukes the ssl directory so that an agent run can reconfigure
just the public and private key before the server is started up.  For
PE, the `ssl_crl_path` is configured at installation time.  The absence
of the CRL file after the test nukes it causes the server to fail to
startup.  Work to be done for SERVER-346 and SERVER-911 would allow for
the CRL file to properly be replaced at startup time during this
scenario.  Since that work has not yet been done, though, this commit
just skips the test when run under PE for now.