(MAINT) Remove 'test/unit/assertions' requires from subcommand tests
Previously, many of the subcommand acceptance tests explicitly required
the 'test/unit/assertions' path.  When running under Ruby 2.3.1 with
only the minitest gem (and not the test-unit gem) installed, this
path does not exist.  This causes the require and, therefore, the tests
to fail.

The minitest gem is pulled in by the beaker gem and defines the
assert_* functions that the tests use.  Rather than either adding the
'test-unit' gem to the Gemfile or changing the tests to require a valid
path in minitest instead, this commit just removes the
'test/unit/assertions' requires since the path should have already been
loaded when running under beaker anyway.  No other tests include require
statements to locate the assert_* functions, so this commit allows the
subcommand tests to follow the same pattern.