(SERVER-1644) Use JAVA_ARGS_CLI env var for Java args in subcommands
Previously, the value of the JAVA_ARGS environment variable, as a result
of work done for the Puppet Server 2.7.0 release, was passed along as
arguments to the Java process running the `gem` subcommand.  This was
problematic for cases where the JAVA_ARGS had been specifically tuned
with arguments for use when running the puppetserver service but
inappropriate for the `gem` command.  For example, Java args that enable
debug logging when included in the `gem` command could break the ability
to perform basic `gem` installs.

This commit changes the `gem` subcommand to source custom Java arguments
from a separate environment variable, `JAVA_ARGS_CLI`.  This commit also
adds the use of the `JAVA_ARGS_CLI` args to the ruby and irb subcommands,
which previously provided no mechanism for specifying custom Java args.