(SERVER-377) Add tests for proxy-related environment variables
Changes to jruby-utils make it so that the environment variables `HTTP_PROXY`,
`http_proxy`, `HTTPS_PROXY`, `https_proxy`, `NO_PROXY`, and `no_proxy` are now
passed through to the jruby scripting container, so that gems can be installed
in scenarios where a proxy is set.

This commit adds acceptance tests ensuring that when set those variables are
available when running `puppetserver ruby` and `puppetserver irb` (and thus
will also be available when running `puppetserver gem`).