(SERVER-1690) Improve CLI subcommand error handling
Previously, if some kinds of errors were thrown during the processing of
a CLI subcommand, e.g., a Prismatic schema error when parsing content
from the puppetserver.conf file, the resulting error displayed on the
command line had no indication of what the real error was.

This commit adds some extra error handling to the `run` function in the
CLI subcommand which, for cases like the schema error, allows the
original upstream exception passed through to the CLI output.  This
would allow the user to see more specific information about the root
cause behind the error.