(PE-19130) Allow puppet conf file to be resources.
Change the puppetserver testutils to allow the sources of the temporary
puppet conf files set up by with-puppet-conf-files to be java.net.URLs
in addition to strings or java.io.Files, and to copy the sources to the
destination files by wrapping the sources in an input stream, which
works even when the source is a URL that points to a classpath resource
contained in a jar.

Together, these changes allow the conf file sources passed to
with-puppet-conf-files to be classpath resources distributed in a jar,
instead of requiring them to be files.