(MAINT) Adjust env-classes tests for gzip in etags
With the previous commit, the bump of tk-jetty9 to 1.7.0 allowed
sufficiently large response bodies to start being properly gzip encoded
again whereas, with even earlier commits, gzip encoding did not work.
When gzip encoding is used in response bodies, the Jetty webserver adds
"--gzip" to the end of the etag header.  The gzip encoding fix caused a
couple of the environment classes tests which had not been expecting the
"--gzip" characters to appear in the etag to start failing.

In this commit, the two failing environment-classes tests were adjusted
to allow for the "--gzip" characters to appear in an etag, allowing the
tests to once again pass.