(SERVER-1723) Use /dev/urandom for gem, ruby, and irb subcommands
This commit adds -Djava.security.egd=/dev/urandom as an argument to the
Java command line for the gem, ruby, and irb subcommands.  This argument
was already being used in the Java command line for when puppetserver
runs as a service.  Performance for some operations like gem installs
appears to be significantly faster with this option in place.