(PE-18344) Add an http-client :metrics-enabled setting
Add a `metrics-enabled` setting to the `http-client` part of the config,
defaulting to true. If ths setting is true, the http client will have a
metrics registry set on it and will use metrics. If the setting is false, the
http client will not set any metrics on requests it makes.