(SERVER-1704) Support flush timeout setting for jruby pool
This commit adds support for a timeout during a jruby pool flush through the
admin api.

Modifies the admin api routes to support returning a 503 if the timeout is
exceeded.

Removes the existing jruby-pool-resource liberator resource and replaces it
with a simpler ring handler. The liberator resource made it prohibitively
difficult to return a 503 Service Unavailable in our use case, and was
unnecessarily complex in general.