(PE-18345) Pass metric-id setting through to http client
This commit makes it so that if a `metric-id` request option is set for a GET
or POST request (the only two request types we support), it is passed through
to the http client's request options. Doing this also meant a small refactor
of some common logic between the `get` and `post` methods, as they both shared
most of the same logic for creating `RequestOptions`.