(SERVER-1260) Move some functions from jruby-metrics service to core ns
This commit moves a couple of functions from the jruby-metrics-service
namespace down into the jruby-metrics-core namespace.  This is being
done to enable the pe-jruby-metrics-service implementation in PE to
potentially reuse this functionality from the core namespace.  A couple
of functions which had been public in the core namespace in order for
the functions in the service space to access them were also made private
since there was no specific need for them to remain public anymore.