(SERVER-1260) Eliminate create-pool-instance with-redef in metrics test
This commit eliminates the use of a with-redef in the
jruby-metrics-service tests around the use of a mock JRubyPool
instance.  The prior code was replaced with the use of existing test
helpers for injecting mock jruby pool instances.  This commit also
reuses some existing helpers for reading the dev bootstrap config, in
place of duplicating the list of services to be bootstrapped.