(SERVER-1262) Reorganized OSS service vars to match PE implementation
This commit reorganizes the initialization of a few of the variables in
the master service implementation in order to more closely match how the
initialization is done in the current PE master service implementation.
This is being done to make it a bit easier to compare the two
implementations going forward.