(SERVER-1262) Initialize SSL files for legacy-ca-routes-disabled test
Previously, if the `legacy-ca-routes-disabled` test was run in isolation
or before any other test in the `legacy-routes-test` namespace, the
server would error out on test startup because no SSL files were in
place.  This commit adds an extra service startup, using the real CA
service, before starting up with the disabled CA service.  This is done
to ensure that all of the necessary SSL files have been created before
the disabled CA service is bootstrapped, allowing the test to be able to
succeed regardless of the order in which it is run.