(SERVER-1262) Separated ca-files-test SSL dirs from master-service-tests
Previously, all of the tests in the master-service-tests namespace which
bootstrap full services shared the same SSL directory configuration.
For the tests which use mock jrubies, the mock jruby puppet
configuration causes different directories to be initialized than for
the non-mock configurations.  This mismatch causes tests to fail if
non-mock and mock tests are intermingled in the same run.

This commit changes the ca-files-test, the one test in the namespace
which intentionally uses non-standard SSL directory locations, to use
its own SSL configuration which is separate from the configuration used
by all of the other tests in the namespace.  This allows for the mixture
of non-mock and mock configurations to result in the same SSL
configuration files being used across all tests, preventing the test
failures which could otherwise be seen.