(SERVER-1744) Move metrics configuration into metrics.conf file
Previously, if a user had modified the puppetserver.conf file after
install of an older package, the new metrics section would not end up
being populated in the puppetserver.conf file after upgrade.  Because
the required server-id key would not be found in the metrics section,
puppetserver would error out on startup.

This commit moves the new metrics configuration from the
puppetserver.conf file to a metrics.conf file in packaging.  This is
being done both for consistency with PE Puppet Server in how the
metrics configuration is presented to users and to allow the required
server-id setting to be populated by default on a package install or
upgrade.