(SERVER-1743) Rework check-for-updates tests for timing issues
Previously, the check-for-updates tests could intermittently fail
because the check-for-updates call can happen later than when assertions
for the call are made.  In this commit, the tests have been reworked to
wait a longer period of time for the call to be made.  To minimize test
run time, the maximum wait time is much longer for the opt-in case vs.
the opt-out case.