(SERVER-1266) Add exclusion for clojure/core.cache in figwheel plugin
Previously, if a lein ezbake task were attempted, it would error out
with a conflict when trying to load clojure.data.priority-map.  This
problem was introduced by the use of the lein-figwheel plugin.  In this
commit, org.clojure/core.cache is excluded for the figwheel plugin,
allowing a version to be pulled in which works properly with lein
ezbake tasks.