(SERVER-1266) Add an auth.conf rule for the dashboard endpoint
This commit adds an auth.conf rule which allows open access to the new
dashboard endpoint.  This commit also adds on a beaker test which
confirms that an attempt to access the dashboard returns an HTTP 200
status code.  The test does not attempt to validate that any content on
the page looks proper.