(SERVER-1753) Add metrics webservice to configuration
This commit adds the metrics webservice to the bootstrap and web routing
configs for both dev and packaging.  This commit includes a simple
beaker test which just confirms that a successful response is received
when the metrics endpoint is queried - basically just ensuring that the
endpoint is registered and that requests aren't being denied by
insufficent authentication.