(SERVER-1264) Add beaker graphite metrics tests
This commit adds a beaker acceptance test, collect_default_metrics,
which sets up a Graphite server, configures Puppet Server to export
metrics to it, and confirms that the default set of metrics that the
server should export are actually exported.  This test is configured to
run on el-6 only because the underlying Puppet module used to setup the
Graphite server, cprice404/grafanadash, only supports el-6 at present.