(SERVER-1737) Add http client metrics to status
Pull out specific http client metrics and add them to the master service
status endpoint under an `http-client-metrics` key. If the http client has not
yet made any requests the value of this key will be nil.

Since other services (specifically, the pe-master service) may have http
client metrics that they would like to add to this list to be included in the
status, the metric ids to be included in the status are registered in an atom
in the service context, and there is a new service function
`add-metric-ids-to-http-client-metrics-list!` that will append metric ids to
this list.