(SERVER-1264) Configure Graphite max_creates_per_min to infinite
Previously, when the collect_default_metrics.rb test was configuring
Graphite, the default Graphite installation was implicitly setting the
MAX_CREATES_PER_MINUTE setting to 50 (by virtue of the default that the
Graphite Puppet module was using).  This caused the total time for
Puppet Server to intially create all metrics to take up to 3 minutes, as
measured when run in the Puppet vmpooler.

This commit adds an extra test step which configures
MAX_CREATES_PER_MINUTE to 'inf' (infinite).  This allows the initial
creation of metrics in the Graphite server to typically take less than 2
seconds, allowing about 3 minutes to be shaved off the run time for the
collect_default_metrics.rb test.