(SERVER-1264) Preserve perms when backing up conf files in metrics test
This commit adds a '-p' flag to the 'cp' commands that the
collect_default_metrics.rb test uses when backing up config files it is
about to change.  Without that flag, puppetserver might not have read
access to the conf files after they are restored, leading to the conf
files failing to be read and the puppetserver service to not start.
With that flag, puppetserver should be able to continue to read the conf
files after they are restored and, therefore, restart properly.