(SERVER-1264) Fix graphite test errors
Previously, the collect_default_metrics test errantly refered to the
variable 'agent', which can be an array of agents, rather than
'graphite', which is a single agent chosen from the array, as the target
for the installation of the graphite server package.  This led to an
error when running the test in a multi-agent configuration.  This commit
fixes the invalid references.