(SERVER-1264) Use proper server when updating graphite in beaker test
This commit corrects the server on which the graphite.pp manifest is
updated for the collect_default_metrics.rb test.  The test needs to
upload the manifest to the server where graphite is being installed but
was errantly uploading it to the master.  In a split configuration, the
master and graphite servers are different and so the test would
previously fail when attempting to upload the file to the wrong server.