(SERVER-1767) Limit concurrent requests to Puppet API
This patch implements a ring handler that rejects incoming requests to the
Puppet API with a 503 status code when the backlog of outstanding requests
exceeds a given limit. The handler can also be configured with a splay limit,
a random fraction of which is set as the value of a Retry-After header that is
returned with the 503 error.