(SERVER-1630) Exclude JRuby dependencies from the uberjar
Previously, core JRuby dependencies from pl/jruby-utils were being built
into the puppetserver uberjar.  This commit removes the core JRuby
dependencies from the uberjar but continuing to pull the removed
dependencies in via a new Maven artifact, pl/jruby-deps.  This is being
done to prepare later on for the use of a separate core JRuby dependency
jar to be supplied on the classpath during packaging.