(SERVER-1630) Added profile for running tasks with jruby9k
This commit adds a profile to the project.clj which pulls in jruby-9k
dependencies, version 9.1.8.0.

This commit also modifies the default `ruby-code-dir` that the server
uses to include a 'classpath:/' prefix.  This prefix is necessary in
order for jruby-9k to be able to properly load from the `ruby-code-dir`.
This prefix was not necessary for jruby-1.7 but does work properly with
jruby-1.7.

This commit also bumps the clj-parent dependency to 0.7.1 in order
to pick up the newer snakeyaml version, 1.18, which is used in JRuby
9.1.8.0 and newer (via the psych gem).