(SERVER-1630) Set default log level for jruby to 'info'
For JRuby 9k in particular, the amount of noise at the DEBUG level
appears to have gotten pretty high - to the point where much of it
doesn't seem very useful.  Adding an entry to the logback.xml to drop
JRuby-related log messages down to the 'info' level by default to reduce
the amount of noise that is added to the logs when setting the default
log level to 'debug'.