(SERVER-1630) Enhanced the excluded jruby jars tests for JRuby 9k
Previously, the `validate-excluded-bc-jars` acceptance test validated
that the Bouncycastle jars in JRuby's packaging did not end up being
repackaged in the Puppet Server release jar.  This was done to avoid
conflicts between Puppet Server's own BC jar dependencies vs. the ones
brought in from JRuby.

For the 9k version of JRuby, a version of the snakeyaml jar is brought
in - in addition to the BC ones.  This commit renames the previous test
to `validate-jruby-jar-deps-excluded-from-uberjars` and validates that
both the BC and snakeyaml jars are not packaged into any of the
uberjars installed from the Puppet Server package.