(SERVER-1788) Use different graphite port in collect_default_metrics
Previously, the collect_default_metrics test would, via the grafanadash
module, try to setup a graphite webserver on port 80.  When run in the
PE Puppet Server Extensions pipeline, this test would fail because
Apache would be unable to listen on port 80 since the PE console already
would have claimed the port.

This commit changes the test to have the graphite web server hosted on
port 50000 in order to avoid the conflict with the PE console.