(SERVER-1630) Added ezbake integration for configurable JRuby jars
This commit alters the ezbake build configuration to produce extra jars
for JRuby 1.7 and 9k which are bundled into puppetserver packages.  This
commit also includes a cli-default script, used for constructing the
JRuby-specific portion of the java CLASSPATH from the value of the
JRUBY_JAR environment variable, and alters the gem, irb, and ruby
subcommand scripts to utilize the cli-default script.  A
'run_with_jruby9k' beaker test was also added for validating that the
puppetserver service and various subcommands can be run with JRuby 9k,
when configured.