(SERVER-1630) Only append classpath:/ to default ruby-code-dir for 9k
Previously, the Ruby load path segment for the Puppet Server's Ruby code
was changed to include a prefix of "classpath:/" so that the code could
be found properly when run under JRuby 9k.  Under JRuby 1.7, however,
including the "classpath:/" prefix broke the puppetserver irb
subcommand.

With this commit, the "classpath:/" prefix is only appended when not
running under JRuby 1.x, compatible with the approach we've used
previously for JRuby 1.7 in Puppet Server while allowing code to load
properly under JRuby 9k.