(SERVER-1630) Additional test cleanup for JRuby 9k
This commit changes a few of the tests which exercise the
`compat-version` setting to match the expected Ruby language version
when run with JRuby 9k. `compat-version` is also made optional in the
test jruby-puppet TK configuration in order to avoid the generation of
warning messages in the test output around the use of a deprecated
setting.