(SERVER-1803) Bump jruby deps and resolve org.ow2.asm conflict
This commit resolves a dependency conflict between JRuby and
org.ow2.asm via several related packaging changes.  Several dependencies
are bumped in this commit:

- jruby-deps (for 1.7) - 1.7.26-1 => 1.7.26-2
- jruby-deps (for 9k) - 9.1.8.0-1 => 9.1.9.0-1
- jruby-utils - 0.9.0 => 0.9.1
- clj-parent - 0.8.0 => 1.1.0
- ezbake - 1.3.0 => 1.4.0

The ezbake changes are needed to build the jruby-deps uberjars properly
along with the changes to resolve the org.ow2.asm conflict.

The new clj-parent version brings in the org.ow2.asm version, 5.0.3, which
resolves the dependency conflict.

The tk-webserver-jetty9 pin is removed because the latest version,
2.0.1, can now be picked up from the newer clj-parent version.