(MAINT) Don't flush non-initialized pool-context at shutdown
Previously, if an exception were thrown before the jruby-puppet-service
had been completely initialized, a stop call to the service on shutdown
would lead to a separate exception being thrown and logged due to trying
to operate on a nil pool-context.

This commit changes the jruby-puppet-service stop logic to avoid trying
to flush the pool-context if it has not yet been created.  This avoids
having the exception be thrown and logged for the nil pool-context.