(SERVER-1661) Bump JRuby dep to 1.7.27 and remove compat-version setting
This commit bumps the jruby-deps dependency for 1.7 from 1.7.26-2 to
1.7.27-1 and jruby-utils dep from 0.9.0 to 0.10.0.

This commit removes support for the jruby-puppet.compat-version setting.
The compat-version setting was introduced in Puppet Server 2.7.0 to
allow for JRuby to run under Ruby language version 2.0 instead of the
default version, 1.9.  Setting compat-version to "2.0" for JRuby 1.7.27,
however, was broken due to a regression - see:
https://github.com/jruby/jruby/issues/4613 - and is unlikely to be fixed
in the future.  For Puppet Server 5+, users will be able to, via the
JRUBY_JAR environment variable in packaging, configure JRuby 9k to be
used instead of JRuby 1.7 for Ruby language version 2+ support.

In conjunction with a corresponding jruby-utils bump, this commit causes
the presence of the jruby-puppet.compat-version setting to throw an
error at startup, with a message indicating how users can configure the
use of JRuby 9k for Ruby language version 2+ support.