(SERVER-1786) Remove site.pp file between Puppet compat tests
Previously, each of the Puppet compat tests uploaded a site.pp file but
did not remove the file as each test completed.  This would lead to any
subsequent tests which did not replace the site.pp file having problems
when an agent run reused the file from the prior test - in particular
the puppetX_version_test.rb.  In this commit, the site.pp file is
cleaned up as part of each test's teardown.